+++ src/main/java/net/frozenblock/wilderwild/mixin/client/wind/fallingleaves/FallingLeafParticleMixin.java
@@ -33,7 +33,7 @@ public class FallingLeafParticleMixin {
 
 	@Unique
 	private static boolean wilderWild$useWind() {
-		return AmbienceAndMiscConfig.get().cloudMovement && ClientWindManager.shouldUseWind();
+		return AmbienceAndMiscConfig.get().wind.cloudMovement && ClientWindManager.shouldUseWind();
 	}
 
 	@ModifyExpressionValue(method = "tick", at = @At(value = "FIELD", target = "Lrandommcsomethin/fallingleaves/util/Wind;windX:F"), require = 0, remap = false)