+++ src/main/java/net/frozenblock/wilderwild/mixin/client/wind/particlerain/SnowFlakeParticleMixin.java
@@ -35,7 +35,7 @@ public class SnowFlakeParticleMixin {
 
 	@Unique
 	private static boolean wilderWild$useWind() {
-		return AmbienceAndMiscConfig.get().cloudMovement && ClientWindManager.shouldUseWind();
+		return AmbienceAndMiscConfig.get().wind.cloudMovement && ClientWindManager.shouldUseWind();
 	}
 
 	@Inject(method = "tick", at = @At("TAIL"), require = 0)