+++ CHANGELOG.md
@@ -71,7 +71,9 @@ Additions
   - This is controlled by the `wilderwild:tumbleweed_passes_through` tag.
 - Wind particles will now occasionally spawn throughout the world.
   - Particles caused by natural Wind can be disabled with the new `Wind Particles` config option in the `Ambience & Misc.` tab.
+    - Added the `Wind Particle Frequency` to control how frequently these will spawn.
   - Particles caused by Wind Disturbances like Breezes and Wind Charges can be disabled with the new `Wind Disturbance Particles` config option in the `Ambience & Misc.` tab.
+    - Added the `Wind Disturbance Particle Frequency` to control how frequently these will spawn.
 
 Bug Fixes, Changes, & Other Additions
 ---
@@ -128,6 +130,8 @@ Bug Fixes, Changes, & Other Additions
   - The minimum value for `Crab Spawn Cap` has been changed to 1 as `Spawn Crabs` makes 0 redundant.
 - Added `block_palm_fronds` to the Mixins config.
 - Renamed the `Misc` config tab to `Ambience & Misc.`
+- Moved the `Water Colors` category from the `Worldgen` tab to the `Ambience & Misc` tab.
+- Added a new `Wind` category in the `Ambience & Misc` tab.
 - Slightly altered the config background screens, now starting at Dirt and ending with Chiseled Mud Bricks instead of starting with Mud and ending with the Stripped Cypress Log.
 
 - Refactored the `BubbleDirection,` `FlowerColor,` and `SlabWillStairSculkBehavior` classes to the `block.impl` package.